refactor: combine frameStyle and frameConfig into frameTheme
