+++ package.json
@@ -46,6 +46,7 @@
     "@sveltejs/vite-plugin-svelte": "^6.2.4",
     "@tsconfig/svelte": "^5.0.8",
     "@types/adm-zip": "^0.5.7",
+    "@types/bun": "^1.3.10",
     "@types/fontkit": "^2.0.8",
     "@types/lodash": "^4.17.23",
     "@types/minimatch": "^5.1.2",
